Сделано улучшение аквариума, добавлен конфиг с параметрами при улучшении.

+++ Assets/_Project/Scripts/Configs/FeederConfig.cs
@@ -8,7 +8,7 @@ public class FeederConfig : ScriptableObject, IValidatableConfig
     public float CooldownSeconds = 8f;
 
     [Tooltip("Общее количество сытости, которое дает одно нажатие на кормушку.")]
-    public float TotalHungerRestorePerUse = 140f;
+    public int TotalHungerRestorePerUse = 140;
 
     [Tooltip("Минимальное и максимальное количество частичек корма за один клик.")]
     [MinMaxSlider(1, 15, true)]
@@ -22,7 +22,7 @@ public class FeederConfig : ScriptableObject, IValidatableConfig
     public void ValidateData()
     {
         CooldownSeconds = Mathf.Max(0.1f, CooldownSeconds);
-        TotalHungerRestorePerUse = Mathf.Max(1f, TotalHungerRestorePerUse);
+        TotalHungerRestorePerUse = Mathf.Max(1, TotalHungerRestorePerUse);
         
         if (FoodPiecesCount.x < 1) FoodPiecesCount.x = 1;
         if (FoodPiecesCount.y < FoodPiecesCount.x) FoodPiecesCount.y = FoodPiecesCount.x;
